requirements.txt
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
conda-pack==0.8.0
contourpy==1.2.0
cycler==0.12.1
einops==0.8.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2024.6.1
h5py==3.8.0
idna==3.4
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.24.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pip==23.0.1
platformdirs==3.8.1
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.30.0
scikit-commpy==0.8.0
scikit-learn==1.4.0
scipy==1.10.1
setuptools==66.0.0
six==1.16.0
snakeviz==2.2.0
sympy==1.13.1
threadpoolctl==3.2.0
tomli==2.0.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchinfo==1.8.0
torchvision==0.20.1+cu118
tornado==6.4
tqdm==4.65.0
triton==3.1.0
typing_extensions==4.9.0
urllib3==2.0.2
wheel==0.38.4
